+++ Murmur3_x64/Murmur3.Testing/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Murmur3.Testing
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}